Switch from nvim-oxi to LuaJIT FFI, fix tree-sitter ABI
- Replace nvim-oxi with plain C FFI (nvim-oxi crashes on Neovim 2.1.x)
- Downgrade tree-sitter to 0.24 to match Neovim's ABI version 15
- Use ffi.load + package.loadlib for reliable native module loading
- Serialize diff results as JSON across FFI boundary
- Token-level diff API working end-to-end:
  diff("function a()...", "function b()...", "lua") returns
  novel tokens with exact line/column positions

+++ src/parse/tree_sitter_converter.rs
@@ -99,7 +99,7 @@ fn build_list_parts<'a>(
 
     // Check if first and last children are matching delimiters
     let first_child = node.child(0);
-    let last_child = node.child(child_count as u32 - 1);
+    let last_child = node.child(child_count - 1);
 
     let mut open_content = String::new();
     let mut open_position = vec![];
